refactor(opencode): improve startup time
