+++ packages/opencode/src/cli/cmd/agent.ts
@@ -2,13 +2,10 @@ import { cmd } from "./cmd"
 import * as prompts from "@clack/prompts"
 import { UI } from "../ui"
 import { Global } from "@opencode-ai/core/global"
-import { Agent } from "../../agent/agent"
-import { Provider } from "@/provider/provider"
 import path from "path"
 import fs from "fs/promises"
 import { Filesystem } from "@/util/filesystem"
 import matter from "gray-matter"
-import { InstanceRef } from "@/effect/instance-ref"
 import { EOL } from "os"
 import type { Argv } from "yargs"
 import { Effect } from "effect"
@@ -62,12 +59,15 @@ const AgentCreateCommand = effectCmd({
         describe: "model to use in the format of provider/model",
       }),
   handler: Effect.fn("Cli.agent.create")(function* (args) {
-    const maybeCtx = yield* InstanceRef
+    const instanceRef = yield* Effect.promise(() => import("@/effect/instance-ref"))
+    const agent = yield* Effect.promise(() => import("../../agent/agent"))
+    const provider = yield* Effect.promise(() => import("@/provider/provider"))
+    const maybeCtx = yield* instanceRef.InstanceRef
     if (!maybeCtx) return yield* Effect.die("InstanceRef not provided")
     const ctx = maybeCtx
-    const agentSvc = yield* Agent.Service
+    const agentSvc = yield* agent.Agent.Service
     const runLocalEffect = <A, E>(effect: Effect.Effect<A, E>) =>
-      Effect.runPromise(effect.pipe(Effect.provideService(InstanceRef, ctx)))
+      Effect.runPromise(effect.pipe(Effect.provideService(instanceRef.InstanceRef, ctx)))
     yield* Effect.promise(async () => {
       const cliPath = args.path
       const cliDescription = args.description
@@ -128,7 +128,7 @@ const AgentCreateCommand = effectCmd({
       // Generate agent
       const spinner = prompts.spinner()
       spinner.start("Generating agent configuration...")
-      const model = args.model ? Provider.parseModel(args.model) : undefined
+      const model = args.model ? provider.Provider.parseModel(args.model) : undefined
       const generated = await runLocalEffect(agentSvc.generate({ description, model })).catch((error) => {
         spinner.stop(`LLM failed to generate agent: ${error.message}`, 1)
         if (isFullyNonInteractive) process.exit(1)
@@ -235,7 +235,8 @@ const AgentListCommand = effectCmd({
   command: "list",
   describe: "list all available agents",
   handler: Effect.fn("Cli.agent.list")(function* () {
-    const agents = yield* Agent.Service.use((svc) => svc.list())
+    const agent = yield* Effect.promise(() => import("../../agent/agent"))
+    const agents = yield* agent.Agent.Service.use((svc) => svc.list())
     const sortedAgents = agents.sort((a, b) => {
       if (a.native !== b.native) {
         return a.native ? -1 : 1
