+++ packages/opencode/src/cli/cmd/debug/index.ts
@@ -3,8 +3,6 @@ import { InstallationVersion } from "@opencode-ai/core/installation/version"
 import { Flag } from "@opencode-ai/core/flag/flag"
 import os from "os"
 import { Duration, Effect } from "effect"
-import { Config } from "@/config/config"
-import { ConfigPlugin } from "@/config/plugin"
 import { effectCmd } from "../../effect-cmd"
 import { cmd } from "../cmd"
 import { ConfigCommand } from "./config"
@@ -52,7 +50,9 @@ const InfoCommand = effectCmd({
   command: "info",
   describe: "show debug information",
   handler: Effect.fn("Cli.debug.info")(function* () {
-    const config = yield* Config.Service.use((cfg) => cfg.get())
+    const configModule = yield* Effect.promise(() => import("@/config/config"))
+    const configPlugin = yield* Effect.promise(() => import("@/config/plugin"))
+    const config = yield* configModule.Config.Service.use((cfg) => cfg.get())
     const termProgram = process.env.TERM_PROGRAM
       ? `${process.env.TERM_PROGRAM}${process.env.TERM_PROGRAM_VERSION ? ` ${process.env.TERM_PROGRAM_VERSION}` : ""}`
       : undefined
@@ -71,7 +71,7 @@ const InfoCommand = effectCmd({
       return
     }
     for (const plugin of config.plugin_origins) {
-      console.log(`- ${ConfigPlugin.pluginSpecifier(plugin.spec)}`)
+      console.log(`- ${configPlugin.ConfigPlugin.pluginSpecifier(plugin.spec)}`)
     }
   }),
 })
