+++ packages/opencode/src/cli/cmd/debug/scrap.ts
@@ -1,16 +1,15 @@
 import { EOL } from "os"
-import { Project } from "@/project/project"
 import * as Log from "@opencode-ai/core/util/log"
-import { makeRuntime } from "@opencode-ai/core/effect/runtime"
 import { cmd } from "../cmd"
-
-const runtime = makeRuntime(Project.Service, Project.defaultLayer)
 
 export const ScrapCommand = cmd({
   command: "scrap",
   describe: "list all known projects",
   builder: (yargs) => yargs,
   async handler() {
+    const project = await import("@/project/project")
+    const { makeRuntime } = await import("@opencode-ai/core/effect/runtime")
+    const runtime = makeRuntime(project.Project.Service, project.Project.defaultLayer)
     const timer = Log.Default.time("scrap")
     const list = await runtime.runPromise((project) => project.list())
     process.stdout.write(JSON.stringify(list, null, 2) + EOL)
