+++ packages/opencode/src/cli/cmd/serve.ts
@@ -1,5 +1,4 @@
 import { Effect } from "effect"
-import { Server } from "../../server/server"
 import { effectCmd } from "../effect-cmd"
 import { withNetworkOptions, resolveNetworkOptions } from "../network"
 import { Flag } from "@opencode-ai/core/flag/flag"
@@ -16,7 +15,7 @@ export const ServeCommand = effectCmd({
       console.log("Warning: OPENCODE_SERVER_PASSWORD is not set; server is unsecured.")
     }
     const opts = yield* resolveNetworkOptions(args)
-    const server = yield* Effect.promise(() => Server.listen(opts))
+    const server = yield* Effect.promise(() => import("../../server/server").then((mod) => mod.Server.listen(opts)))
     console.log(`opencode server listening on http://${server.hostname}:${server.port}`)
 
     yield* Effect.never
