+++ packages/opencode/src/cli/cmd/web.ts
@@ -1,5 +1,4 @@
 import { Effect } from "effect"
-import { Server } from "../../server/server"
 import { UI } from "../ui"
 import { effectCmd } from "../effect-cmd"
 import { withNetworkOptions, resolveNetworkOptions } from "../network"
@@ -41,7 +40,7 @@ export const WebCommand = effectCmd({
       UI.println(UI.Style.TEXT_WARNING_BOLD + "!  OPENCODE_SERVER_PASSWORD is not set; server is unsecured.")
     }
     const opts = yield* resolveNetworkOptions(args)
-    const server = yield* Effect.promise(() => Server.listen(opts))
+    const server = yield* Effect.promise(() => import("../../server/server").then((mod) => mod.Server.listen(opts)))
     UI.empty()
     UI.println(UI.logo("  "))
     UI.empty()
