+++ packages/opencode/src/cli/effect-cmd.ts
@@ -1,8 +1,7 @@
 import type { Argv } from "yargs"
 import { Effect, Schema } from "effect"
-import { AppRuntime, type AppServices } from "@/effect/app-runtime"
-import { InstanceStore } from "@/project/instance-store"
-import { InstanceRef } from "@/effect/instance-ref"
+import type { AppServices } from "@/effect/app-runtime"
+import type { InstanceStore } from "@/project/instance-store"
 import { cmd, type WithDoubleDash } from "./cmd/cmd"
 
 /**
@@ -74,21 +73,26 @@ export const effectCmd = <Args, A>(opts: EffectCmdOpts<Args, A>) =>
     describe: opts.describe,
     builder: opts.builder as never,
     async handler(rawArgs) {
+      const runtime = await import("@/effect/app-runtime")
       // yargs typing wraps Args in ArgumentsCamelCase<WithDoubleDash<...>>; cast at the boundary.
       const args = rawArgs as unknown as WithDoubleDash<Args>
       const useInstance = typeof opts.instance === "function" ? opts.instance(args) : opts.instance !== false
       if (!useInstance) {
-        await AppRuntime.runPromise(opts.handler(args))
+        await runtime.AppRuntime.runPromise(opts.handler(args))
         return
       }
+      const instanceStore = await import("@/project/instance-store")
+      const instanceRef = await import("@/effect/instance-ref")
       const directory = opts.directory?.(args) ?? process.cwd()
-      const { store, ctx } = await AppRuntime.runPromise(
-        InstanceStore.Service.use((store) => store.load({ directory }).pipe(Effect.map((ctx) => ({ store, ctx })))),
+      const { store, ctx } = await runtime.AppRuntime.runPromise(
+        instanceStore.InstanceStore.Service.use((store) =>
+          store.load({ directory }).pipe(Effect.map((ctx) => ({ store, ctx }))),
+        ),
       )
       try {
-        await AppRuntime.runPromise(opts.handler(args).pipe(Effect.provideService(InstanceRef, ctx)))
+        await runtime.AppRuntime.runPromise(opts.handler(args).pipe(Effect.provideService(instanceRef.InstanceRef, ctx)))
       } finally {
-        await AppRuntime.runPromise(store.dispose(ctx))
+        await runtime.AppRuntime.runPromise(store.dispose(ctx))
       }
     },
   })
